+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/MetadataLogFileCatalog.scala
@@ -25,7 +25,11 @@ import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.execution.datasources._
 
 
-class StreamFileCatalog(sparkSession: SparkSession, path: Path)
+/**
+ * A [[FileCatalog]] that generates the list of files to processing by reading them from the
+ * metadata log files generated by the [[FileStreamSink]].
+ */
+class MetadataLogFileCatalog(sparkSession: SparkSession, path: Path)
   extends PartitioningAwareFileCatalog(sparkSession, Map.empty, None) {
 
   private val metadataDirectory = new Path(path, FileStreamSink.metadataDir)
